+++ src/BreadServiceProvider.php
@@ -4,6 +4,9 @@ namespace Libaro\Bread;
 
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Support\ServiceProvider;
+use Libaro\Bread\Commands\CreateCustomField;
+use Libaro\Bread\Commands\CreateCustomFilter;
+use Libaro\Bread\Commands\CreateCustomHeader;
 use Libaro\Bread\Filters\Filters;
 use Libaro\Bread\Services\FilterService;
 use Libaro\Bread\Services\SortingService;
@@ -35,7 +38,9 @@ class BreadServiceProvider extends ServiceProvider
     private function registerCommands(): void
     {
         $this->commands([
-            // Add commands here
+            CreateCustomField::class,
+            CreateCustomHeader::class,
+            CreateCustomFilter::class,
         ]);
     }