+++ src/Renderers/IndexRenderer.php
@@ -7,14 +7,12 @@ namespace Libaro\Bread\Renderers;
 use Libaro\Bread\Contracts\Renderer;
 use Libaro\Bread\Filters\Filters;
 use Libaro\Bread\Headers\Headers;
-use Libaro\Bread\Routes\Routes;
 
 final class IndexRenderer extends Renderer
 {
     protected string $title = '';
     protected ?Headers $headers = null;
     protected ?Filters $filters = null;
-    protected ?Routes $routes = null;
     protected $actions = [];
     protected $items = [];