+++ CHANGELOG.md
@@ -1,4 +1,30 @@
 # Changelog
 
 # 1.0.0
+
 - Initial release
+
+## v1.3.1 - 2022-07-07
+
+Adds scope support to filter service
+
+## v1.3.0 - 2022-07-06
+
+- Adds support for custom filters
+- Adds command to create custom Filter
+
+## v1.2.0 - 2022-07-06
+
+Adds command to create custom Header
+
+## v1.1.1 - 2022-07-06
+
+Removes version from composer.json
+
+## v1.1.0 - 2022-07-06
+
+Adds command to create custom Field
+
+## v1.0.0 - 2022-07-05
+
+Initial release